requirements.txt
allure-pytest==2.8.18
allure-python-commons==2.8.18
atomicwrites==1.4.0
attrs==20.2.0
autoit==0.1
certifi==2020.6.20
chardet==3.0.4
clipboard==0.0.4
colorama==0.4.3
configparser==5.0.0
crayons==0.4.0
enum34==1.1.10
idna==2.10
iniconfig==1.0.1
lxml==4.5.2
more-itertools==8.5.0
namedlist==1.8
packaging==20.4
Pillow==7.2.0
pluggy==0.13.1
py==1.9.0
PyAutoIt==0.6.0
PyHamcrest==2.0.2
pyparsing==2.4.7
pyperclip==1.8.0
pytest==6.1.0
requests==2.24.0
selenium==3.141.0
six==1.15.0
SSIM-PIL==1.0.10
toml==0.10.1
urllib3==1.25.10
vk-api==11.9.0
wcwidth==0.2.5
webdriver-manager==3.2.2
